requirements.txt
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
idna==3.4
lxml==4.9.3
ncclient==0.6.13
paramiko==3.2.0
pyaml==23.7.0
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
urllib3==2.0.3
